Create piet crate Move windows specific code platform directory
