+++ piet/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "piet"
+version = "0.1.0"
+
+[dependencies]
+directwrite = "0.1.2"
+direct2d = "0.2.0"
