+++ piet/src/windows/brush.rs
@@ -1,0 +1,1 @@
+pub use direct2d::brush::*;
