+++ piet/src/windows/render_target.rs
@@ -1,0 +1,1 @@
+pub use direct2d::render_target::*;
