+++ src/widget/key_listener.rs
@@ -14,8 +14,7 @@
 
 //! Widget for forwarding key events to a listener.
 
-use winapi::um::winuser::*;
-use druid_win_shell::window::M_ALT;
+use druid_shell::window::*;
 
 use widget::Widget;
 use {HandlerCtx, Id, KeyEvent, KeyVariant, Ui};
